added login and user auth part; cleaned deprecated code;

+++ app/routes.py
@@ -2,20 +2,21 @@
 from werkzeug.utils import secure_filename
 
 from flask import current_app as app
-from flask import make_response, render_template, request
-import deprecation
-from wtforms import ValidationError
+from flask import make_response, render_template, request, redirect, url_for, flash
 import uuid, shortuuid
 import sys, os.path
 from app import app_url, app_port
+from flask_login import current_user, login_required, login_user, logout_user
 
-from .models import Movie, db, MoviePath, MovieInfo
-from .forms import RegisterForm, MoviePathForm, MovieInfoForm
+from .models import db, MoviePath, MovieInfo, User
+from .forms import MoviePathForm, MovieInfoForm, LoginForm, RegistrationForm
 
 url = "http://"+app_url+":"+app_port
 
 @app.route("/movie/", methods=["GET"])
 def __index():
+    if not current_user.is_authenticated:
+        return redirect(url_for('loginUser'))
     allMovies = MovieInfo.query.all()
 
     if allMovies:
@@ -26,6 +27,8 @@ def __index():
 
 @app.route("/movie/<short>", methods=["GET"])
 def index(short):
+    if not current_user.is_authenticated:
+        return redirect(url_for('loginUser'))
     existing = MoviePath.query.filter(MoviePath.uuid == short).first()
 
     if existing:
@@ -33,41 +36,10 @@ def index(short):
     else:
         return make_response("There's no movie for '{}' existing. ".format(short))
 
-@app.route("/", methods=["GET"])
-@deprecation.deprecated(deprecated_in="1.x", removed_in="2.0",
-                        current_version="1.0",
-                        details="Use 'RegisterMovie()' instead.")
-def addMovie():
-    nameEN = request.args.get("nameEN")
-    nameCN = request.args.get("nameCN")
-    path = request.args.get("path")
-    fileName = request.args.get("filename")
-    short = request.args.get("short")
-
-    try:
-        newMovie = validateAndAddMovie(nameEN, nameCN, path, fileName, short)
-        return render_template("home/index.html", file=newMovie.filePath(), allMovies=Movie.query.all(), url=url)
-    except ValidationError as ve:
-        return make_response("Add new Movie failed because: {}".format(ve))
-    except Exception as e:
-        return make_response("Add new Movie failed and don't know why. Detail: {}".format(e))
-
-@app.route("/register", methods=['GET', 'POST'])
-def registerMovie():
-    form = RegisterForm()
-    if form.validate_on_submit():
-        fn = secure_filename(form.file.data)
-        try:
-            newMovie = validateAndAddMovie(form.nameEN.data, form.nameCN.data, form.path.data, fn, form.short.data)
-            return render_template("home/index.html", file=newMovie.filePath(), allMovies=Movie.query.all(), url=url)
-        except ValidationError as ve:
-            return make_response("Add new Movie failed because: {}".format(ve))
-        except Exception as e:
-            return make_response("Add new Movie failed and don't know why. Detail: {}".format(e))
-    return render_template("home/register.html", form=form)
-
 @app.route("/addAll", methods=['GET', 'POST'])
 def addAllMovive():
+    if not current_user.is_authenticated:
+        return redirect(url_for('loginUser'))
     form = MoviePathForm()
     if form.validate_on_submit():
         if os.path.isdir(os.path.join(sys.path[0], "app", "static", form.path.data)):
@@ -78,6 +50,8 @@ def addAllMovive():
 
 @app.route("/movie/edit/<uuid>", methods=['GET', 'POST'])
 def editMoviveInfo(uuid):
+    if not current_user.is_authenticated:
+        return redirect(url_for('loginUser'))
     path = MoviePath.query.filter(MoviePath.uuid == uuid).first()
     old = MovieInfo.query.filter(MovieInfo.uuid == uuid).first()
     infoForm = MovieInfoForm()
@@ -92,6 +66,8 @@ def editMoviveInfo(uuid):
 
 @app.route("/movie/delete/<uuid>", methods=['GET', 'POST'])
 def deleteMovie(uuid):
+    if not current_user.is_authenticated:
+        return redirect(url_for('loginUser'))
     MoviePath.query.filter(MoviePath.uuid == uuid).delete()
     MovieInfo.query.filter(MovieInfo.uuid == uuid).delete()
     db.session.commit()
@@ -120,22 +96,38 @@ def secureFileName(dir):
                 db.session.add(info)
                 db.session.commit()
 
-def validateAndAddMovie(nameEN: str, nameCN: str, path: str, filename: str, short: str):
-    if nameEN is None:
-        raise ValidationError("Name in EN must not be null. ")
-    if short is None:
-        raise ValidationError("Short Name must not be null. ")
 
-    existName = Movie.query.filter(Movie.nameEN == nameEN).first()
-    existShort = Movie.query.filter(Movie.short == short).first()
+@app.route("/login", methods=['GET', 'POST'])
+def loginUser():
+    if current_user.is_authenticated:
+        return redirect(url_for('__index'))
+    form = LoginForm()
+    if form.validate_on_submit():
+        user = User.query.filter_by(username=form.username.data).first()
+        if user is None or not user.check_password(form.password.data):
+            flash('Invalid username or password')
+            return redirect(url_for('loginUser'))
+        login_user(user, remember=form.remember_me.data)
+        return redirect(url_for('__index'))
+    return render_template('home/login.html', title='Sign In', form=form)
 
-    if existName:
-        raise ValidationError("File: {} already exists, adding movie failed. ".format(nameEN))
-    if existShort:
-        raise ValidationError("ShortName: {} already exists. Can not use it to add new movie. ".format(short))
+@app.route('/logout')
+def logoutUser():
+    logout_user()
+    return redirect(url_for('loginUser'))
 
-    newMovie = Movie(nameEN=nameEN, nameCN=nameCN, path=path, fileName=filename, short=short)
-    db.session.add(newMovie)
-    db.session.commit()
-    return newMovie
+@app.route('/register', methods=['GET', 'POST'])
+def register():
+    if current_user.is_authenticated:
+        return redirect(url_for('__index'))
+    form = RegistrationForm()
+    if form.validate_on_submit():
+        user = User(username=form.username.data, email=form.email.data)
+        user.set_password(form.password.data)
+        db.session.add(user)
+        db.session.commit()
+        flash('Congratulations, you are now a registered user!')
+        return redirect(url_for('loginUser'))
+    return render_template('home/register.html', title='Register', form=form)
+
